Feature/BL-64 Implement DTO, mappers, repositories for Buildings and Flats [BE]

* BL-64 add BuildingDto implementation, BuildingMapper and BuildingRepository

* BL-64 correction Manager implementation

* BL-64 add response status to CRUD methods, BuildingController, BuildingService, BuildingProvider

* BL-64 add FlatDto implementation, FlatRepository,FlatMapper

* BL-64 add FlatService, FlatProvider, FlatController

* BL-64 add mock data to database and chang in application.properties (spring.sql.init.mode=always chang to spring.datasource.initialization-mode=always)

* BL-64 complete announcement, building, manager and member implementations

* BL-64 add mock data to building, flat, member and announcement


Approved-by: Przemyslaw Szafranski

+++ src/main/java/com/capgemini/bedland/manager/internal/ManagerServiceImpl.java
@@ -1,13 +1,12 @@
 package com.capgemini.bedland.manager.internal;
 
+import com.capgemini.bedland.exceptions.NotFoundException;
 import com.capgemini.bedland.manager.api.ManagerEntity;
 import com.capgemini.bedland.manager.api.ManagerProvider;
-import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Service
 class ManagerServiceImpl implements ManagerService, ManagerProvider {
@@ -20,17 +19,20 @@ class ManagerServiceImpl implements ManagerService, ManagerProvider {
 
     @Override
     public List<ManagerDto> getAll() {
-        return managerMapper.entities2Dtos(managerRepository.findAll());
+        return managerMapper.entities2DTOs(managerRepository.findAll());
     }
 
     @Override
     public ManagerDto getById(final Long id) {
-        return managerMapper.entity2Dto(managerRepository.findById(id).get());
+        return managerMapper.entity2Dto(managerRepository.findById(id)
+                                                         .orElseThrow(() -> new NotFoundException(id)));
     }
 
     @Override
     public ManagerDto create(final ManagerDto request) {
-        if(request.getId() != null) throw new IllegalArgumentException("Given request contains an ID. Manager can't be created");
+        if (request.getId() != null) {
+            throw new IllegalArgumentException("Given request contains an ID. Manager can't be created");
+        }
         ManagerEntity newManager = managerMapper.dto2Entity(request);
         ManagerEntity createdManager = managerRepository.save(newManager);
         return managerMapper.entity2Dto(createdManager);
@@ -38,23 +40,21 @@ class ManagerServiceImpl implements ManagerService, ManagerProvider {
 
     @Override
     public ManagerDto update(final ManagerDto request) {
-        if(request.getId() == null) throw new IllegalArgumentException("Given request has no ID");
-        if(!managerRepository.existsById(request.getId())) {
-            throw new NoSuchElementException("Manager with ID = " + request.getId() + " does not exist");
-            // TODO: add error codes (404)
+        if (request.getId() == null) {
+            throw new IllegalArgumentException("Given request has no ID");
         }
-        ManagerEntity updatedManager = managerRepository.findById(request.getId()).orElseThrow();
-        updatedManager.setLogin(request.getLogin());
-        updatedManager.setName(request.getName());
-        updatedManager.setLastName(request.getLastName());
-        updatedManager.setEmail(request.getEmail());
-        updatedManager.setPassword(request.getPassword());
-        updatedManager.setPhoneNumber(request.getPhoneNumber());
+        if (!managerRepository.existsById(request.getId())) {
+            throw new NotFoundException(request.getId());
+        }
+        ManagerEntity updatedManager = managerMapper.dto2Entity(request);
         return managerMapper.entity2Dto(managerRepository.save(updatedManager));
     }
 
     @Override
     public void delete(final Long id) {
+        if (!managerRepository.existsById(id)) {
+            throw new NotFoundException(id);
+        }
         managerRepository.deleteById(id);
     }
